fixed norm issues
