+++ philo/src/utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anakin <anakin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: apregitz <apregitz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/05 14:37:03 by apregitz          #+#    #+#             */
-/*   Updated: 2025/09/20 17:13:02 by anakin           ###   ########.fr       */
+/*   Updated: 2025/09/21 15:06:36 by apregitz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,16 +32,40 @@ int ft_strcmp(char *s1, char *s2)
 
 void	print_status(t_philo *philo, char *status)
 {
-	long long	timestamp;
-	bool		death_check;
+	bool	death_check;
+	int		death_msg;
 
+	death_msg = ft_strcmp(status, "died");
+	if (death_msg != 0)
+		pthread_mutex_lock(&philo->data->print_mutex);
 	pthread_mutex_lock(&philo->death_mutex);
 	death_check = philo->dead;
 	pthread_mutex_unlock(&philo->death_mutex);
-	if (death_check)
-		return;
-	timestamp = get_time_in_ms() - philo->data->start_time;
-	pthread_mutex_lock(&philo->data->print_mutex);
-	printf("%lld %d %s\n", timestamp, philo->id, status);
-	pthread_mutex_unlock(&philo->data->print_mutex);
+	if (death_msg != 0 && death_check)
+		return ((void)pthread_mutex_unlock(&philo->data->print_mutex));
+	printf("%ld %d %s\n", get_time_in_ms() - philo->data->start_time, philo->id, status);
+	if (death_msg != 0)
+		pthread_mutex_unlock(&philo->data->print_mutex);
+}
+
+int	ft_usleep(long time_in_ms, t_philo *philo)
+{
+	long	start_time;
+	long	elapsed;
+	bool	death_check;
+
+	start_time = get_time_in_ms();
+	while (1)
+	{
+		elapsed = get_time_in_ms() - start_time;
+		pthread_mutex_lock(&philo->death_mutex);
+		death_check = philo->dead;
+		pthread_mutex_unlock(&philo->death_mutex);
+		if (death_check)
+			return (1);
+		if (elapsed >= time_in_ms)
+			break ;
+		usleep(100);
+	}
+	return (0);
 }
